# src/GestMantIA.API/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "GestMantIA API",
    "description": "API para el sistema de Gestión de Mantenimiento Inteligente Asistido.",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "Soporte GestMantIA",
      "url": "https://gestmantia.com/contact",
      "email": "soporte@gestmantia.com"
    },
    "license": {
      "name": "Licencia de Uso",
      "url": "https://example.com/license"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/users/{userId}/lock": {
      "post": {
        "tags": [
          "UserLockout"
        ],
        "summary": "Bloquea a un usuario por un período de tiempo específico.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID del usuario a bloquear.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "durationMinutes",
            "in": "query",
            "description": "Duración del bloqueo en minutos (opcional, si no se especifica, el bloqueo es indefinido).",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "description": "Razón del bloqueo (opcional).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{userId}/unlock": {
      "post": {
        "tags": [
          "UserLockout"
        ],
        "summary": "Desbloquea a un usuario previamente bloqueado.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID del usuario a desbloquear.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{userId}/lockout-info": {
      "get": {
        "tags": [
          "UserLockout"
        ],
        "summary": "Obtiene la información de bloqueo de un usuario.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID del usuario.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserLockoutInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserLockoutInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserLockoutInfo"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "UserLockoutInfo": {
        "required": [
          "UserName"
        ],
        "type": "object",
        "properties": {
          "UserId": {
            "type": "string",
            "format": "uuid"
          },
          "UserName": {
            "type": "string",
            "nullable": true
          },
          "IsLockedOut": {
            "type": "boolean"
          },
          "LockoutEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "Reason": {
            "type": "string",
            "nullable": true
          },
          "LockoutStart": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "IsPermanent": {
            "type": "boolean"
          },
          "TimeRemaining": {
            "type": "string",
            "nullable": true,
            "readOnly": true,
            "example": "00:00:00"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Encabezado de autorización JWT usando el esquema Bearer. Ejemplo: \"Authorization: Bearer {token}\"",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}